+++ araqne-logdb/src/main/java/org/araqne/logdb/query/command/OutputTxt.java
@@ -199,7 +199,7 @@ public class OutputTxt extends QueryCommand {
 	@Override
 	public void onClose(QueryStopReason reason) {
 		close();
-		if (reason == QueryStopReason.CommandFailure)
+		if (!append && reason == QueryStopReason.CommandFailure)
 			if (tmpPath != null)
 				new File(tmpPath).delete();
 			else
